Completed add trial fragments

+++ MogulMoves/app/src/main/java/com/example/mogulmoves/ExperimentSerializer.java
@@ -3,6 +3,8 @@ package com.example.mogulmoves;
 import java.util.HashMap;
 import java.util.List;
 
+import static java.sql.Types.NULL;
+
 /**
  * Class to convert Experiment objects into savable data and vice-versa.
  */
@@ -68,8 +70,8 @@ public class ExperimentSerializer implements Serializer<Experiment> {
         int owner = (int) (long) map.get("owner");
         int id = (int) (long) map.get("id");
 
-        List<Integer> trials = (List<Integer>) map.get("trials");
-        List<Integer> messages = (List<Integer>) map.get("messages");
+        List<Long> trials = (List<Long>) map.get("trials");
+        List<Long> messages = (List<Long>) map.get("messages");
 
         if(type == 0) {
             experiment = new BinomialExperiment(id, owner, description, region,
@@ -91,15 +93,15 @@ public class ExperimentSerializer implements Serializer<Experiment> {
         experiment.setActive(active);
 
         try {
-            for(int trial: trials){
-                experiment.addTrial(trial);
+            for(long trial: trials){
+                experiment.addTrial((int) trial);
             }
         } catch (java.lang.NullPointerException e) {
         }
 
         try {
-            for (int message : messages) {
-                experiment.addMessage(message);
+            for (long message : messages) {
+                experiment.addMessage((int) message);
             }
         } catch (java.lang.NullPointerException e) {
         }
